Adding Dummy Observations

This page notes how to define dummy observation matrices in Bayesian VAR and VHAR models.

Bayesian VAR

Consider BVAR and its hyperparameters in set_bvar().

sigma: \(\sigma_1, \ldots, \sigma_m\)

lambda: \(\lambda\)

delta: \(\delta_1, \ldots, \delta_m\)

eps: \(\epsilon\)

This package implements adding-dummy-observations approach of Bańbura et al. (2010). Let \(J_p = diag(1, 2, \ldots, p)\). For each response and design matrix \(Y_0\) and \(X_0\), we define dummy observation named by \(Y_p\) and \(X_p\). Response dummy matrix \(Y_p\) is (m + k) x m matrix: $$ Y_p = \left[\begin{array}{c} diag\left( \delta_1 \sigma_1, \ldots, \delta_m \sigma_m \right) / \lambda \\ 0_{m(p - 1) \times m} \\ \hline diag\left( \sigma_1, \ldots, \sigma_m \right) \\ \hline 0_m^\intercal \end{array}\right] $$ Design dummy matrix \(X_p\) is (m + k) x k matrix: $$ X_p = \left[\begin{array}{c|c} J_p \otimes diag\left( \sigma_1, \ldots, \sigma_m \right) / \lambda & 0_{mp} \\ \hline 0_{m \times mp} & 0_m \\ \hline 0_{mp}^\intercal & \epsilon \end{array}\right] $$

These two matrices define Minnesota prior distribution of BVAR.

Bayesian VHAR

Consider BVHAR and its hyperparameter in set_bvhar(). First, VAR-type minnesota prior:

sigma: \(\sigma_1, \ldots, \sigma_m\)

lambda: \(\lambda\)

delta: \(\delta_1, \ldots, \delta_m\)

eps: \(\epsilon\) For response matrix \(Y_0\), define (m + h) x m matrix \(Y_{HAR}\) $$ Y_{HAR} = \left[\begin{array}{c} diag\left( \delta_1 \sigma_1, \ldots, \delta_m \sigma_m \right) / \lambda \\ 0_{2m \times m} \\ \hline diag\left( \sigma_1, \ldots, \sigma_m \right) \\ \hline 0_m^\intercal \end{array}\right] $$ For design matrix \(X_0\), define (m + h) x h matrix \(X_{HAR}\) $$ X_{HAR} = \left[\begin{array}{c|c} J_3 \otimes diag\left( \sigma_1, \ldots, \sigma_m \right) / \lambda & 0_{3m} \\ \hline 0_{m \times 3m} & 0_m \\ \hline 0_{3m}^\intercal & \epsilon \end{array}\right] $$ In case of VHAR-type minnesota prior, delta is replaced with the following three:

daily: \(d_1, \ldots, d_m\)

weekly: \(w_1, \ldots, w_m\)

monthly: \(m_1, \ldots, m_m\) and \(Y_{HAR}\) is changed. $$ Y_{HAR} = \left[\begin{array}{c} diag\left( d_1 \sigma_1, \ldots, d_m \sigma_m \right) / \lambda \\ diag\left( w_1 \sigma_1, \ldots, w_m \sigma_m \right) / \lambda \\ diag\left( m_1 \sigma_1, \ldots, m_m \sigma_m \right) / \lambda \\ \hline diag\left( \sigma_1, \ldots, \sigma_m \right) \\ \hline 0_m^\intercal \end{array}\right] $$

These two dummy matrices define minnesota prior distribution of BVHAR.

References

Bańbura, M., Giannone, D., & Reichlin, L. (2010). Large Bayesian vector auto regressions. Journal of Applied Econometrics, 25(1).