Fitting Bayesian VAR-SV

This function fits VAR-SV. It can have Minnesota, SSVS, and Horseshoe prior.

Usage

bvar_sv(
  y,
  p,
  num_iter = 1000,
  num_burn = floor(num_iter/2),
  thinning = 1,
  bayes_spec = set_bvar(),
  include_mean = TRUE,
  minnesota = FALSE,
  verbose = FALSE,
  num_thread = 1
)

# S3 method for bvarsv
print(x, digits = max(3L, getOption("digits") - 3L), ...)

knit_print.bvarsv(x, ...)

Arguments

y

Time series data of which columns indicate the variables

p

VAR lag

num_iter

MCMC iteration number

num_burn

Number of burn-in (warm-up). Half of the iteration is the default choice.

thinning

Thinning every thinning-th iteration

bayes_spec

A BVAR model specification by set_bvar().

include_mean

Add constant term (Default: TRUE) or not (FALSE)

minnesota

Apply cross-variable shrinkage structure (Minnesota-way). By default, FALSE.

verbose

Print the progress bar in the console. By default, FALSE.

num_thread

Number of threads

x

bvarsv object

digits

digit option to print

...

not used

Value

bvar_sv() returns an object named bvarsv

class.

Details

Cholesky stochastic volatility modeling for VAR based on $$\Sigma_t = L^T D_t^{-1} L$$

References

Carriero, A., Chan, J., Clark, T. E., & Marcellino, M. (2022). Corrigendum to “Large Bayesian vector autoregressions with stochastic volatility and non-conjugate priors” [J. Econometrics 212 (1)(2019) 137–154]. Journal of Econometrics, 227(2), 506-512.

Chan, J., Koop, G., Poirier, D., & Tobias, J. (2019). Bayesian Econometric Methods (2nd ed., Econometric Exercises). Cambridge: Cambridge University Press.

Cogley, T., & Sargent, T. J. (2005). Drifts and volatilities: monetary policies and outcomes in the post WWII US. Review of Economic Dynamics, 8(2), 262–302.

Gruber, L., & Kastner, G. (2022). Forecasting macroeconomic data with Bayesian VARs: Sparse or dense? It depends! arXiv.