Extract Log-Likelihood of Multivariate Time Series Model

Compute log-likelihood function value of VAR(p), VHAR, BVAR(p), and BVHAR

Usage

# S3 method for class 'varlse'
logLik(object, ...)

# S3 method for class 'vharlse'
logLik(object, ...)

# S3 method for class 'bvarmn'
logLik(object, ...)

# S3 method for class 'bvarflat'
logLik(object, ...)

# S3 method for class 'bvharmn'
logLik(object, ...)

Arguments

object

Model fit

...

not used

Value

A logLik object.

Details

Consider the response matrix \(Y_0\). Let \(n\) be the total number of sample, let \(m\) be the dimension of the time series, let \(p\) be the order of the model, and let \(s = n - p\). Likelihood of VAR(p) has

$$Y_0 \mid B, \Sigma_e \sim MN(X_0 B, I_s, \Sigma_e)$$

where \(X_0\) is the design matrix, and MN is matrix normal distribution.

Then log-likelihood of vector autoregressive model family is specified by

$$\log p(Y_0 \mid B, \Sigma_e) = - \frac{sm}{2} \log 2\pi - \frac{s}{2} \log \det \Sigma_e - \frac{1}{2} tr( (Y_0 - X_0 B) \Sigma_e^{-1} (Y_0 - X_0 B)^T )$$

In addition, recall that the OLS estimator for the matrix coefficient matrix is the same as MLE under the Gaussian assumption. MLE for \(\Sigma_e\) has different denominator, \(s\).

$$\hat{B} = \hat{B}^{LS} = \hat{B}^{ML} = (X_0^T X_0)^{-1} X_0^T Y_0$$ $$\hat\Sigma_e = \frac{1}{s - k} (Y_0 - X_0 \hat{B})^T (Y_0 - X_0 \hat{B})$$ $$\tilde\Sigma_e = \frac{1}{s} (Y_0 - X_0 \hat{B})^T (Y_0 - X_0 \hat{B}) = \frac{s - k}{s} \hat\Sigma_e$$

In case of VHAR, just consider the linear relationship.

While frequentist models use OLS and MLE for coefficient and covariance matrices, Bayesian models implement posterior means.

References

Lütkepohl, H. (2007). New Introduction to Multiple Time Series Analysis. Springer Publishing.

Corsi, F. (2008). A Simple Approximate Long-Memory Model of Realized Volatility. Journal of Financial Econometrics, 7(2), 174–196.

Bańbura, M., Giannone, D., & Reichlin, L. (2010). Large Bayesian vector auto regressions. Journal of Applied Econometrics, 25(1).

Litterman, R. B. (1986). Forecasting with Bayesian Vector Autoregressions: Five Years of Experience. Journal of Business & Economic Statistics, 4(1), 25.

Ghosh, S., Khare, K., & Michailidis, G. (2018). High-Dimensional Posterior Consistency in Bayesian Vector Autoregressive Models. Journal of the American Statistical Association, 114(526).

See also