Oxford-Man Institute Realized Library

The realized measure of financial assets dataset provided by Oxford-man Institute of Quantitative Finance.

Usage

oxfordman_rv

oxfordman_rk

Format

oxfordman_long is the raw data frame of 53507 rows and 20 columns (You cannot call this dataset.):

date

Date - From 2012-01-09 to 2015-06-27

Symbol

Name of the Assets - See below for each name

nobs

Number of observations

by_ss

Bipower Variation (5-min Sub-sampled)

rsv

Realized Semi-variance (5-min)

rk_parzen

Realized Kernel Variance (Non-Flat Parzen)

rv10

Realized Variance (10-min)

rv5_ss

Realized Variance (5-min Sub-sampled)

rv5

Realized Variance (5-min)

rv10_ss

Realized Variance (10-min Sub-sampled)

rk_twoscale

Realized Kernel Variance (Two-Scale/Bartlett)

close_price

Closing (Last) Price

rsv_ss

Realized Semi-variance (5-min Sub-sampled)

rk_th2

Realized Kernel Variance (Tukey-Hanning(2))

open_time

Opening Time

medrv

Median Realized Variance (5-min)

open_price

Opening (First) Price

bv

Bipower Variation (5-min)

open_to_close

Open to Close Return

close_time

Closing Time

oxfordman_rv is a data frame that interpolates NA values of oxfordman_wide_rv. Also, it does not have date column for fitting. The number of rows is 905 and the number of columns is 30 (except date).

date

Date - From 2012-01-09 to 2015-06-27

AEX

AEX index

AORD

All Ordinaries

BFX

Bell 20 Index

BSESN

S&P BSE Sensex

BVLG

PSI All-Share Index (excluded because this index is observed from 2012-10-15)

BVSP

BVSP BOVESPA Index

DJI

Dow Jones Industrial Average

FCHI

CAC 40

FTMIB

FTSE MIB

FTSE

FTSE 100

GDAXI

DAX

GSPTSE

S&P/TSX Composite index

HSI

HANG SENG Index

IBEX

IBEX 35 Index

IXIC

Nasdaq 100

KS11

Korea Composite Stock Price Index (KOSPI)

KSE

Karachi SE 100 Index

MXX

IPC Mexico

N225

Nikkei 225

NSEI

NIFTY 50

OMXC20

OMX Copenhagen 20 Index

OMXHPI

OMX Helsinki All Share Index

OMXSPI

OMX Stockholm All Share Index

OSEAX

Oslo Exchange All-share Index

RUT

Russel 2000

SMSI

Madrid General Index

SPX

S&P 500 Index

SSEC

Shanghai Composite Index

SSMI

Swiss Stock Market Index

STI

Straits Times Index (excluded because this index is NA in the period)

STOXX50E

EURO STOXX 50

oxfordman_rk is a data frame that interpolates NA values of oxfordman_wide_rk. Also, it does not have DATE column for fitting. The number of rows is 1826 and the number of columns is 31.

Source

Realized library of oxford-man had been discontinued, so the source could not be listed.

Details

As a raw dataset, we have internal dataset of long format oxfordman_long. It contains every realized measure.

Denote that non-trading dates are excluded in oxfordman_long, not in NA. So be careful when dealing this set directly.

For analysis, we widened the data for 5-min realized volatility (rv5) and realized kernel variance (rk_parzen), respectively.

oxfordman_wide_rv

oxfordman_wide_rk

oxford_rv and oxford_rk are the sets whose NA values interpolated using imputeTS::na_interpolation().

First three datasets should be called using data() function: data(..., package = "bvhar").

Only oxford_rv and oxford_rk is lazy loaded.