Stochastic Search Variable Selection in VAR

This page describes a stochastic search variable selection (SSVS) MCMC algorithm in a VAR model.

SSVS Prior

Consider the vectorized formulation \(vec(Y_0) = (I_k \otimes X_0) vec(A) + vec(Z_0)\). As the other Bayesian VAR model, this model handles coefficients \(A\) and variance matrix \(\Sigma_e\). To shrink \(\Sigma_e^{-1}\), however, upper cholesky factor is used as the alternative \(\Sigma_e^{-1} = \Psi \Psi^\intercal\) in this context.

Prior of coefficients

Each \(vec(A) = \alpha = (\alpha_1, \ldots, \alpha_{k^2 p + k})\) except constant-corresponding term is restricted by \(\gamma = (\gamma_1, \ldots, \gamma_{k^2 p})^\intercal\), which is dummy vector. Then $$ h_i = \begin{cases} \tau_{0i} & \text{if } \gamma_i = 0 \\ \tau_{1i} & \text{if } \gamma_i = 1 \end{cases} $$ with small \(\tau_{0j}\) and large \(\tau_{1j}\). In turn, \(D = diag(h_1, \ldots, h_{k^2 p})\).

Let \(\alpha_{coef}\) be the restricted coefficients vector and let \(\alpha_{non}\) be the not-restricted coefficients vector, i.e. vectorized constant term. Each term has its own prior. $$ \alpha_{coef} \mid \gamma \sim N(0_{k^2 p}, DD), \quad \alpha_{non} \sim N(0_k, c I_k) $$ If \(c\) is large, then prior influence to \(\alpha_{non}\) decreases. By combining each term in appropriate order, $$\alpha \mid \gamma \sim N(0_{k^2 p + k}, M)$$ is acquired by $$ M_0 = I_{k p + 1} \otimes \begin{bmatrix} DD & 0_{k^2 p} \\ 0_{k^2 p}^\intercal & c \end{bmatrix} $$ Sometimes nonzero prior mean \(\alpha_0\) is also considered.

Prior of Coefficients Restrictions

It is natural that 0-1 \(\gamma_{j}\) has Bernoulli distribution. $$\gamma_j \sim Bernoulli(p_j)$$ If there is no information, set \(p_j = 0.5\).

Prior of Cholesky Factor

Let \(\Psi = [\psi_{ij}] \in \mathbb{R}^{k \times k}\). Recall that \(\Psi\) is an upper triangular matrix such that \(\Sigma_e^{-1} = \Psi \Psi^\intercal\).

To define the prior distribution, George et al. (2008) divide the matrix in two parts

Diagnonal element: \(\psi = (\psi_{11}, \ldots, \psi_{kk})^\intercal\)

Off-diagonal element: \(\eta_j = (\psi_{1j}, \ldots, \psi_{j-1, j})^\intercal, j = 2, \ldots, k\)

Prior of off-diagonal element

To restrict cholesky factor, dummy vector \(\omega_j = (\omega_{1j}, \ldots, \omega_{j - 1, j})^\intercal\) corresponding to off-diagonal elements are defined. Then the matrix \(D_j = diag(h_{1j}, \ldots, h_{j-1, j})\) is defined by $$ h_{ij} = \begin{cases} \kappa_{0ij} & \text{if } \omega_{ij} = 0 \\ \kappa_{1ij} & \text{if } \omega_{ij} = 1 \end{cases} $$ with small \(\kappa_{0ij}\) and large \(\kappa_{1ij}\). As coefficients vector, \(\eta_j\) has normal distribution $$\eta_j \mid \omega_j \sim N(0_{j - 1}, D_j D_j), j = 2, \ldots, k$$

Prior of off-diagonal element restrictions

As \(\gamma_j\), \(\omega_{ij}\) has Bernoulli distribution. $$\omega_{ij} \sim Bernoulli(q_{ij})$$ \(q_{ij} = 0.5\) is the most common choice.

Prior of diagonal element

Since cholesky factor is positive definite, diagonal element of the matrix is given Gamma distribution. $$\psi_{jj}^2 \sim Gamma(shape = a_j, rate = b_j)$$

Gibbs Sampling

Full Conditional

George et al. (2008) presents every full conditionals. Let SSE matrix be \(S(\hat{A}) = (Y_0 - X_0 \hat{A})^\intercal (Y_0 - X_0 \hat{A}) \in \mathbb{R}^{k \times k}\), let \(S_j\) be the upper-left j x j block matrix of \(S(A)\), and let \(s_j = (s_{1j}, \ldots, s_{j - 1, j})^\intercal\) of \(S(A)\).

For specified shape and rate of Gamma distribution \(a_j\) and \(b_j\), $$ \psi_{jj}^2 \mid \alpha, \gamma, \omega, Y_0 \sim \gamma(a_i + n / 2, B_i) $$ where $$ B_i = \begin{cases} b_1 + s_{11} / 2 & \text{if } i = 1 \\ b_i + (s_{ii} - s_i^\intercal ( S_{i - 1} + (D_i R_i D_i)^(-1) )^(-1) s_i) & \text{if } i = 2, \ldots, k \end{cases} $$ , and \(D_i = diag(h_{1j}, \ldots, h_{i - 1, i}) \in \mathbb{R}^{(j - 1) \times (j - 1)}\).

For every \(j = 1, \ldots, k\), $$ \eta_j \mid \alpha, \gamma, \omega, \psi, Y_0 \sim N (\mu_j, \Delta_j) $$ where $$ \mu_j = -\psi_{jj} (S_{j - 1} + (D_j R_j D_j)^{-1})^{-1} s_j \in \mathbb{R}^{j - 1}, \Delta_j = (S_{j - 1} + (D_j R_j D_j)^{-1})^{-1} \in \mathbb{R}^{(j - 1) \times (j - 1)} $$

Consider restriction dummy vectors. $$ \omega_{ij} \mid \eta_j, \psi_j, \alpha, \gamma, \omega_{j}^{(previous)} \sim Bernoulli(\frac{u_{ij1}}{u_{ij1} + u_{ij2}}) $$ where $$ u_{ij1} = \frac{1}{\kappa_{1ij}} \exp(- \frac{\psi_{ij}^2}{2 \kappa_{1ij}^2}) q_{ij}, u_{ij2} = \frac{1}{\kappa_{0ij}} \exp(- \frac{\psi_{ij}^2}{2 \kappa_{0ij}^2}) (1 - q_{ij}) $$ Also, $$ \gamma_j \mid \alpha, \psi, \eta, \omega, Y_0 \sim Bernoulli(\frac{u_{i1}}{u_{j1} + u_{j2}}) $$ where $$ u_{j1} = \frac{1}{\tau_{1j}} \exp(- \frac{\alpha_j^2}{2 \tau_{1j}^2})p_i, u_{j2} = \frac{1}{\tau_{0j}} \exp(- \frac{\alpha_j^2}{2 \tau_{0j}^2})(1 - p_i) $$

In case of coefficients vector, $$\alpha \mid \gamma, \eta, \omega, \psi, Y_0 \sim N (\mu, \Delta)$$ where $$ \mu = ( ' (\Psi \Psi^\intercal) \otimes (X_0 X_0^\intercal) + M^{-1} )^{-1} ( ' ( (\Psi \Psi^\intercal) \otimes (X_0 X_0^\intercal) ) \hat{\alpha} + M^{-1} \alpha_0 ) $$ where \(\alpha_0\) is the prior mean for \(\alpha \mid \gamma\), and \(\hat{\alpha}\) is MLE (equivalently, OLS). $$ \Delta = ((\Psi \Psi^\intercal) \otimes (X_0 X_0^\intercal) + M^{-1})^{-1} $$

Gibbs Sampling

Data: \(X_0\), \(Y_0\)

Input:

VAR order p

MCMC iteration number

Weight of each slab: Bernoulli distribution parameters

\(p_j\): of coefficients

\(q_{ij}\): of cholesky factor

Gamma distribution parameters for cholesky factor diagonal elements \(\psi_{jj}\)

\(a_j\): shape

\(b_j\): rate

Correlation matrix of coefficient vector: \(R = I_{k^2p}\)

Correlation matrix to restrict cholesky factor (of \(\eta_j\)): \(R_j = I_{j - 1}\)

Tuning parameters for spike-and-slab sd semi-automatic approach

\(c_0\): small value (0.1)

\(c_1\): large value (10)

Constant to reduce prior influence on constant term: \(c\)

Algorithm:

Initialize \(\Psi\), \(\omega\), \(\alpha\), \(\gamma\)

Iterate

Diagonal elements of cholesky factor: \(\psi^{(t)} \mid \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Off-diagonal elements of cholesky factor: \(\eta^{(t)} \mid \psi^{(t)} \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Dummy vector for cholesky factor: \(\omega^{(t)} \mid \eta^{(t)}, \psi^{(t)} \alpha^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Coefficient vector: \(\alpha^{(t)} \mid \gamma^{(t - 1)}, \Sigma^{(t)}, \omega^{(t)}, Y_0\)

Dummy vector for coefficient vector: \(\gamma^{(t)} \mid \alpha^{(t)}, \psi^{(t)}, \eta^{(t)}, \omega^{(t)}, Y_0\)

Output:

Parameter trace

Update results

OLS

References

George, E. I., Sun, D., & Ni, S. (2008). Bayesian stochastic search for VAR model restrictions. Journal of Econometrics, 142(1), 553–580. doi:10.1016/j.jeconom.2007.08.017

Koop, G., & Korobilis, D. (2009). Bayesian Multivariate Time Series Methods for Empirical Macroeconomics. Foundations and Trends® in Econometrics, 3(4), 267–358. doi:10.1561/0800000013