Stochastic Search Variable Selection in VHAR

This page describes a stochastic search variable selection (SSVS) MCMC algorithm in a VHAR model. Recall that \(\Sigma_e = \Psi \Psi^\intercal\).

SSVS Prior

Prior of coefficients

Among \(vec(\Phi) = \phi = (\phi_1, \ldots, \phi_{3 k^2 + k})\), non-constant terms are restricted by dummy vector \(\gamma = (\gamma_1, \ldots, \gamma_{3 k^2})^\intercal\). This defines the diagonal matrix \(D = diag(h_1, \ldots, h_{k^2 p})\) by $$ h_i = \begin{cases} \tau_{0i} & \text{if } \gamma_i = 0 \\ \tau_{1i} & \text{if } \gamma_i = 1 \end{cases} $$ with small \(\tau_{0j}\) and large \(\tau_{1j}\).

Let \(\phi_{coef}\) be the restricted coefficients vector and let \(\phi_{non}\) be the not-restricted coefficients vector, i.e. vectorized constant term. Each term has its own prior. $$ \phi_{coef} \mid \gamma \sim N(\phi_{0, coef}, DD), \quad \alpha_{non} \sim N(\phi_{0, non}, c I_k) $$ If \(c\) is large, then prior influence to \(\phi_{non}\) decreases. By combining each term in appropriate order, $$\phi \mid \gamma \sim N(0_{3 k^2 + k}, M)$$ with $$ M = I_{k p + 1} \otimes \begin{bmatrix} DD & 0_{k^2 p} \\ 0_{k^2 p}^\intercal & c \end{bmatrix} $$ and \(\phi_0\) combined in the same way.

Prior of Other Parameters

We are using the the notations for the other parameters, so see ssvs_bvar_algo.

Gibbs Sampling

Data: \(X_0\), \(Y_0\), VAR linear transformation

Input:

VHAR order (week, month)

MCMC iteration number

Weight of each slab: Bernoulli distribution parameters

\(p_j\): of coefficients

\(q_{ij}\): of cholesky factor

Gamma distribution parameters for cholesky factor diagonal elements \(\psi_{jj}\)

\(a_j\): shape

\(b_j\): rate

Correlation matrix of coefficient vector: \(R = I_{k^2p}\)

Correlation matrix to restrict cholesky factor (of \(\eta_j\)): \(R_j = I_{j - 1}\)

Tuning parameters for spike-and-slab sd semi-automatic approach

\(c_0\): small value (0.1)

\(c_1\): large value (10)

Constant to reduce prior influence on constant term: \(c\)

Algorithm:

Initialize \(\Psi\), \(\omega\), \(\phi\), \(\gamma\)

Iterate

Diagonal elements of cholesky factor: \(\psi^{(t)} \mid \phi^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Off-diagonal elements of cholesky factor: \(\eta^{(t)} \mid \psi^{(t)} \phi^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Dummy vector for cholesky factor: \(\omega^{(t)} \mid \eta^{(t)}, \psi^{(t)} \phi^{(t - 1)}, \gamma^{(t - 1)}, \omega^{(t - 1)}, Y_0\)

Coefficient vector: \(\phi^{(t)} \mid \gamma^{(t - 1)}, \Sigma^{(t)}, \omega^{(t)}, Y_0 \sim \mu, \Delta)\)

where \(\mu = ((\Psi \Psi^\intercal) \otimes (X_1 X_1^\intercal) + M^{-1})^{-1} (( (\Psi \Psi^\intercal) \otimes (X_1 X_1^\intercal) ) \hat{\phi} + M^{-1} \phi_0)\)

and \(\Delta = ((\Psi \Psi^\intercal) \otimes (X_1 X_1^\intercal) + M^{-1})^{-1}\)

Dummy vector for coefficient vector: \(\gamma^{(t)} \mid \phi^{(t)}, \psi^{(t)}, \eta^{(t)}, \omega^{(t)}, Y_0\)

Output:

Parameter trace

Update results

OLS

References

George, E. I., Sun, D., & Ni, S. (2008). Bayesian stochastic search for VAR model restrictions. Journal of Econometrics, 142(1), 553–580.

Koop, G., & Korobilis, D. (2009). Bayesian Multivariate Time Series Methods for Empirical Macroeconomics. Foundations and Trends® in Econometrics, 3(4), 267–358.