bvhar: Bayesian Vector Heterogeneous Autoregressive Modeling

Tools to model and forecast multivariate time series including Bayesian Vector heterogeneous autoregressive (VHAR) model by Kim & Baek (2023) (doi:10.1080/00949655.2023.2281644 ). 'bvhar' can model Vector Autoregressive (VAR), VHAR, Bayesian VAR (BVAR), and Bayesian VHAR (BVHAR) models.

Details

The bvhar package provides function to analyze and forecast multivariate time series data via vector autoregressive modeling. Here, vector autoregressive modelling includes:

Vector autoregressive (VAR) model: var_lm()

Vector heterogeneous autoregressive (VHAR) model: vhar_lm()

Bayesian VAR (BVAR) model: var_bayes()

Bayesian VHAR (BVHAR) model: vhar_bayes()

References

Kim, Y. G., and Baek, C. (2024). Bayesian vector heterogeneous autoregressive modeling. Journal of Statistical Computation and Simulation, 94(6), 1139-1157.

Kim, Y. G., and Baek, C. (n.d.). Working paper.

See also

Author

Maintainer: Young Geun Kim ygeunkimstat@gmail.com (ORCID) [copyright holder]

Other contributors:

Changryong Baek [contributor]