Finding the Set of Hyperparameters of Bayesian Model

This function chooses the set of hyperparameters of Bayesian model using stats::optim() function.

Usage

choose_bayes(
  bayes_bound = bound_bvhar(),
  ...,
  eps = 1e-04,
  y,
  order = c(5, 22),
  include_mean = TRUE,
  parallel = list()
)

Arguments

bayes_bound

Empirical Bayes optimization bound specification defined by bound_bvhar().

...

Additional arguments for stats::optim().

eps

Hyperparameter eps is fixed. By default, 1e-04.

y

Time series data

order

Order for BVAR or BVHAR. p of bvar_minnesota() or har of bvhar_minnesota(). By default, c(5, 22) for har.

include_mean

Add constant term (Default: TRUE) or not (FALSE)

parallel

List the same argument of optimParallel::optimParallel(). By default, this is empty, and the function does not execute parallel computation.

Value

bvharemp

class is a list that has

...

Many components of stats::optim() or optimParallel::optimParallel()

spec

Corresponding bvharspec

fit

Chosen Bayesian model

ml

Marginal likelihood of the final model

References

Giannone, D., Lenza, M., & Primiceri, G. E. (2015). Prior Selection for Vector Autoregressions. Review of Economics and Statistics, 97(2). doi:10.1162/REST_a_00483

Kim, Y. G., and Baek, C. (2022). Bayesian vector heterogeneous autoregressive modeling. submitted.

See also